Add hash_full_python_code option to ReturnnConfig (#118)
As discussed in #93, function bodies are not hashed when passed to a ReturnnConfig in python_prolog or python_epilog. We agreed not to change that since it would break all hashes. The option hash_full_python_code in the ReturnnConfig now  allows hashing the parsed python_prolog or python_epilog, i.e., a change to a function body would change the hash.