Output WER to relpath instead of to abspath
When `prefix=""`, the output was being written to `/wers/...`, which is undesirable. With this change, we ensure that the output is always correctly written to the relative path `output/wers/...`. Note that for `prefix` to work correctly it needs to be a directory, i.e. it needs to finish with `/`, but I think that's a precondition that's already set two lines above, when setting more outputs to `{prefix}recog_...`