Add functionality to merge subcorpora and recordings for MergeCorporaJob (#496)
`MERGE_RECURSIVE` creates an empty corpus, and treats the corpora passed as parameters as if they had to be actually merged as one. It does so by adding the recordings and the subcorpora (recursively) to the parent corpus. The result is a single corpus with all speakers combined and with all segments of a given recording merged.

Using this flag could cover use cases for parallelizing Bliss processing jobs through a split-segments + process + merge-recursive approach.

Besides, I added a bit of documentation concerning the merging strategies, as well as tests concerning all merging strategies.