ReturnnTrainingJob info, more robust LR log loading
In my case, I had some np.float64(...) entries in it
(for the effective LR).
Not sure how I got that (maybe due to new PyTorch/Numpy),
and I already pushed some RETURNN change to avoid this in the future,
but I think we can anyway just handle that here as well.