Fix newline in Bliss corpus orth
The actual functionality of the code seems ambiguous with respect to the explanatory comment some lines above.

In order to keep the information from the writing, the newline should be substituted by a space, and not by an empty string. Example:

```
HELLO
WORLD
```

Current functionality: `HELLOWORLD`.

My proposal is consistent with the human way of reading: `HELLO WORLD`.